requirements.txt
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-RESTful==0.3.9
requests==2.31.0
mysql-connector-python==8.1.0
